requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
twilio==9.2.2
stripe==12.3.0
redis==5.0.7
pytest==8.3.2
python-dotenv==1.0.1
openai==1.96.0
tenacity==8.5.0
Faker==26.0.0
freezegun==1.5.1
python-dateutil==2.9.0
amadeus==12.0.0
cloudinary
fpdf2==2.8.3
psycopg2-binary==2.9.10
pytz==2024.1
pypdf==4.2.0
pytest-mock==3.14.1
geopy==2.4.1
timezonefinder==6.5.9
web3==7.13.0
eth-account==0.13.7